pyproject.toml
[project]
name = "catgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.4.10",
    "langdetect>=1.0.9",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "parquet>=1.3.1",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.4",
    "streamlit>=1.40",
    "joblib>=1.3",
    "perplexityai>=0.22.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
